requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
Pillow==10.4.0
httpx==0.27.0
supabase==2.9.1
